# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: API Instituto
  version: 1.0.0
  description: |
    API REST del Instituto.

    ### Forma de las respuestas

    **En exito el body es el recurso pelado**, sin ningun sobre: un objeto en
    las respuestas de un recurso y un array en los listados. Un listado vacio
    devuelve `200 []`, nunca un 404. Un `DELETE` devuelve `204` sin body.

    **En error** el body es un objeto `{ success: false, message, details? }`,
    igual para todos los endpoints porque sale del manejador centralizado
    (`src/middlewares/error.middleware.js`).

    Son formas distintas a proposito: para saber si algo salio bien mirá el
    **status HTTP**, no una propiedad del body.

    Las respuestas de error estan definidas una sola vez en
    `components/responses`: no las redefinas, usalas con `$ref`.
servers:
  - url: https://api-instituto.onrender.com/api
    description: Produccion
  - url: http://localhost:3000/api
    description: Desarrollo local

# Un tag por recurso. Agrupa los endpoints en la UI y define el orden en que
# aparecen.
...
paths:
...
  /sedes/{id}:
    get:
      tags: [Sedes]
      security:
        # Auth opcional: el endpoint es publico, pero si llega un token se usa
        # (lo necesita `incluirInactivas`). El `{}` es lo que lo marca como
        # opcional; sin la lista entera, Swagger UI NO manda el token aunque
        # hayas apretado Authorize, y el request sale anonimo.
        - {}
        - bearerAuth: []
      summary: Obtiene una sede por id, con sus carreras
      description: |
        A diferencia del listado, el detalle incluye las carreras que se
        dictan en la sede.

        Una sede **dada de baja devuelve 404**, igual que si no existiera. Con
        `incluirInactivas=true` se devuelve de todos modos, pero eso exige un
        token de ADMIN.

        El endpoint en si es publico.
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la sede
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
        - name: incluirInactivas
          in: query
          required: false
          description: >-
            Con `true` incluye tambien las sedes dadas de baja. **Requiere token
            con rol ADMIN**: sin el, la respuesta es 403. Por defecto quedan
            invisibles. Solo se aceptan los valores `true` y `false`; cualquier
            otro da 400.
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
      responses:
        "200":
          description: La sede pedida, con sus carreras
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SedeDetalle" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [Sedes]
      summary: Modifica una sede
      description: |
        Actualizacion **parcial**: se manda solo lo que cambia. Hay que enviar
        al menos un campo, un body vacio es un 400.

        Los tres casos que conviene tener claros:

        | Lo que mandas | Que pasa |
        |---|---|
        | omitis el campo | queda como estaba |
        | `"telefono": null` | se borra el valor |
        | `"telefono": ""` | **400** — para borrar se usa `null` |

        `nombre`, `ciudad` y `provincia` son obligatorios en la base: no se
        pueden poner en `null`, aunque si cambiar por otro valor.

        Los textos se recortan (trim) antes de guardarse. La respuesta es la
        sede sola, **sin** las carreras.

        **No se puede editar una sede dada de baja**: devuelve 404. Primero hay
        que reactivarla.

        **Requiere token con rol ADMIN.**
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la sede
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SedeUpdateInput" }
      responses:
        "200":
          description: La sede ya modificada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sede" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >-
            El cambio dejaria dos sedes con el mismo nombre en la misma ciudad.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [Sedes]
      summary: Da de baja una sede
      description: |
        **Baja logica**: la sede no se borra de la base, se marca con
        `activa: false`. Sus vinculos con carreras se conservan, que es lo que
        permite reactivarla despues sin perder nada.

        A partir de la baja deja de aparecer en el listado y en el detalle,
        salvo que se pida `incluirInactivas=true`.

        Sobre una sede **que ya estaba dada de baja devuelve 404**, no 204: es
        la contracara de que una sede dada de baja sea invisible.

        **Requiere token con rol ADMIN.**
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la sede
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "204":
          description: Dada de baja. Sin body.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: La sede no existe o ya estaba dada de baja.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /publicaciones/{id}:
    get:
      tags: [Publicaciones]
      summary: Obtiene una publicación por id
      description: |
        Endpoint público. Incluye los datos del autor (sin información sensible).
        Si no existe devuelve 404.
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la publicación
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "200":
          description: La publicación pedida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publicacion" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [Publicaciones]
      summary: Modifica una publicación
      description: |
        Actualización **parcial**: se manda solo lo que cambia. Hay que enviar
        al menos un campo, un body vacío es 400.

        Campos modificables: `titulo`, `resumen`, `contenido`, `tipo`, `imagenUrl`, `fechaEvento`, `destacada`.
        No se permite modificar manualmente `id`, `slug`, `autorId`, `createdAt`, `updatedAt`.

        El `slug` permanece estable aunque cambie el `titulo`, para no romper URLs.

        Para `destacada` basta con mandar `{"destacada": true}` o `false` en el PATCH normal;
        no hay endpoint especial.

        Contrato de nulos igual que Sedes:
        | Lo que mandás | Qué pasa |
        |---|---|
        | omitís el campo | queda como estaba |
        | `"resumen": null` | se borra el valor |
        | `"resumen": ""` | **400** — para borrar se usa `null` |

        Regla `fechaEvento`/`tipo` se valida sobre el **estado final** del registro
        (combinando valores actuales con los nuevos). Por eso un PATCH que cambie
        solo `tipo` a `EVENTO` sin `fechaEvento` dará 400 si el registro no tenía fecha,
        y cambiar a `NOTICIA` fuerza `fechaEvento` a `null`.

        **Requiere token con rol ADMIN.**
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la publicación
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicacionUpdateInput" }
      responses:
        "200":
          description: La publicación ya modificada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Publicacion" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [Publicaciones]
      summary: Elimina una publicación
      description: |
        Eliminación **física**: la fila se borra de la base (no hay `activa` ni `deletedAt` en este modelo).
        Si no existe devuelve 404. Si se elimina correctamente devuelve 204 sin body.

        **Requiere token con rol ADMIN.**
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: cuid de la publicación
          schema: { type: string }
          example: clx1a2b3c4d5e6f7g8h9i0j1
      responses:
        "204":
          description: Eliminada. Sin body.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    # Es lo que valida `authenticate` en src/middlewares/auth.middleware.js.
    # Para probar endpoints protegidos desde la UI: boton "Authorize" y pegar
    # el token SIN el prefijo "Bearer " (la UI lo agrega sola).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Sede:
      type: object
      description: Una sede del instituto.
      required: [id, nombre, ciudad, provincia, activa, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: cuid generado por la base.
          examples: ["clx1a2b3c4d5e6f7g8h9i0j1"]
        nombre:
          type: string
          examples: ["Sede Centro"]
        ciudad:
          type: string
          examples: ["Cordoba"]
        provincia:
          type: string
          examples: ["Cordoba"]
        direccion:
          type: [string, "null"]
          examples: ["Av. Colon 1234"]
        telefono:
          type: [string, "null"]
          examples: ["+54 351 4567890"]
        email:
          type: [string, "null"]
          format: email
          examples: ["centro@instituto.edu.ar"]
        activa:
          type: boolean
          description: >-
            `false` significa que la sede esta dada de baja. Por defecto esas
            sedes no aparecen ni en el listado ni en el detalle.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SedeUpdateInput:
      type: object
      description: >-
        Body del PATCH. Todos los campos son opcionales, pero hay que mandar al
        menos uno. Los tres opcionales aceptan `null` para borrar el valor. No
        acepta campos que no esten en esta lista.
      minProperties: 1
      additionalProperties: false
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 120
          examples: ["Sede Centro II"]
        ciudad:
          type: string
          minLength: 1
          maxLength: 80
          examples: ["Cordoba"]
        provincia:
          type: string
          minLength: 1
          maxLength: 80
          examples: ["Cordoba"]
        direccion:
          type: [string, "null"]
          minLength: 1
          maxLength: 200
          description: "`null` borra la direccion."
          examples: ["Av. Colon 1234"]
        telefono:
          type: [string, "null"]
          minLength: 1
          maxLength: 30
          description: "`null` borra el telefono."
          examples: ["+54 351 4567890"]
        email:
          type: [string, "null"]
          format: email
          maxLength: 120
          description: "`null` borra el email."
          examples: ["centro@instituto.edu.ar"]
    SedeDetalle:
      description: |
        Una sede con las carreras que se dictan en ella. Lo devuelve
        `GET /sedes/{id}`.

        Dos cosas a tener en cuenta:

        - Cada elemento de `carreras` viene **envuelto** en un objeto
          `{ carrera: ... }`. No es un array de carreras: es la tabla
          intermedia de la relacion N:M. Del lado del front:
          `sede.carreras.map(c => c.carrera)`.
        - **Incluye las carreras dadas de baja**, marcadas con
          `activa: false`. Filtrarlas es decision del front.
      allOf:
        - $ref: "#/components/schemas/Sede"
        - type: object
          required: [carreras]
          properties:
            carreras:
              type: array
              description: Ordenadas por nombre de la carrera.
              items:
                type: object
                required: [carrera]
                properties:
                  carrera: { $ref: "#/components/schemas/CarreraResumen" }
...
    CarreraResumen:
      type: object
      description: >-
        Version reducida de una Carrera, para cuando viaja anidada dentro de
        otro recurso. No trae `descripcion` ni los timestamps.
      required: [id, nombre, slug, modalidad, activa]
      properties:
        id:
          type: string
          examples: ["clc1a2b3c4d5e6f7g8h9i0j1"]
        nombre:
          type: string
          examples: ["Tecnicatura Superior en Programacion"]
        slug:
          type: string
          examples: ["tecnicatura-superior-en-programacion"]
        modalidad:
          type: string
          enum: [PRESENCIAL, VIRTUAL, HIBRIDA]
        duracionAnios:
          type: [integer, "null"]
          examples: [3]
        tituloOtorgado:
          type: [string, "null"]
          examples: ["Tecnico Superior en Programacion"]
        activa:
          type: boolean
          description: >-
            `false` significa que la carrera esta dada de baja. Viene igual en
            la respuesta.
...
    Publicacion:
      type: object
      description: Noticia o evento del instituto.
      required: [id, titulo, slug, contenido, tipo, destacada, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: cuid generado por la base.
          examples: ["clx1a2b3c4d5e6f7g8h9i0j1"]
        titulo:
          type: string
          examples: ["Jornada de Inteligencia Artificial 2026"]
        slug:
          type: string
          description: Slug único generado a partir del título. Permanece estable aunque cambie el título.
          examples: ["jornada-de-inteligencia-artificial-2026"]
        resumen:
          type: [string, "null"]
          examples: ["Una jornada abierta para estudiantes."]
        contenido:
          type: string
          examples: ["Contenido completo de la noticia..."]
        tipo:
          type: string
          enum: [NOTICIA, EVENTO]
          examples: ["EVENTO"]
        imagenUrl:
          type: [string, "null"]
          format: uri
          examples: ["https://example.com/imagen.jpg"]
        fechaEvento:
          type: [string, "null"]
          format: date-time
          description: Solo para EVENTO. En NOTICIA siempre es null.
          examples: ["2026-10-20T18:00:00.000Z"]
        destacada:
          type: boolean
          description: Si es destacada aparece primero o resaltada en el front.
          examples: [true]
        autorId:
          type: [string, "null"]
          description: CUID del autor (usuario que creó la publicación). Null si fue eliminado.
        autor:
          type: ["object", "null"]
          description: Datos del autor sin información sensible. Solo id, nombre y apellido.
          properties:
            id:
              type: string
            nombre:
              type: string
            apellido:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PublicacionUpdateInput:
      type: object
      description: >-
        Body del PATCH. Todos los campos son opcionales, pero hay que mandar al menos uno.
        resumen e imagenUrl aceptan `null` para borrar el valor. No acepta campos que no estén en esta lista.
        Regla tipo/fechaEvento se valida sobre el estado final: EVENTO requiere fechaEvento, NOTICIA fuerza null.
      minProperties: 1
      additionalProperties: false
      properties:
        titulo:
          type: string
          minLength: 1
          maxLength: 200
          examples: ["Jornada de Inteligencia Artificial 2026 - Actualizado"]
        resumen:
          type: [string, "null"]
          minLength: 1
          maxLength: 500
          description: "`null` borra el resumen."
          examples: ["Resumen actualizado."]
        contenido:
          type: string
          minLength: 1
          examples: ["Contenido actualizado..."]
        tipo:
          type: string
          enum: [NOTICIA, EVENTO]
          examples: ["NOTICIA"]
        imagenUrl:
          type: [string, "null"]
          format: uri
          maxLength: 500
          description: "`null` borra la imagen."
          examples: ["https://example.com/nueva-imagen.jpg"]
        fechaEvento:
          type: [string, "null"]
          format: date-time
          description: "`null` borra la fecha (solo válido si el estado final es NOTICIA)."
          examples: ["2026-11-15T18:00:00.000Z"]
        destacada:
          type: boolean
          examples: [false]
...
    Error:
      type: object
      description: Forma de cualquier respuesta de error de la API.
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          description: Descripcion del error, apta para mostrar al usuario.
          examples: ["El registro solicitado no existe"]
        details:
          description: Informacion extra. Solo aparece en algunos errores.
        stack:
          type: string
          description: >-
            Traza del error. SOLO se incluye cuando NODE_ENV=development;
            en produccion nunca se expone.
    ValidationIssue:
      type: object
      description: Un campo que no paso la validacion de Zod.
      required: [source, field, message]
      properties:
        source:
          type: string
          enum: [body, params, query]
          description: En que parte del request estaba el campo.
        field:
          type: string
          examples: ["email"]
        message:
          type: string
          examples: ["Invalid email address"]
    ValidationError:
      type: object
      description: >-
        Error 400 devuelto por el middleware `validate()`. A diferencia del
        resto, `details` trae la lista de campos que fallaron.
      required: [success, message, details]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
          examples: ["Datos de entrada invalidos"]
        details:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
  responses:
...
    BadRequest:
      description: Los datos enviados no pasaron la validacion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Unauthorized:
      description: Falta el token, es invalido, expiro o la cuenta esta desactivada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Autenticado, pero el rol no alcanza para esta accion.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: El recurso no existe.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"